requirements.txt
requests
python-dotenv
aiogram
apscheduler
